requirements.txt
Flask
gunicorn
SQLAlchemy>=1.4.46,<=1.4.46
Flask-SQLAlchemy
pymysql
python-dotenv 
Flask-Script